feat: add interfaces for writing PCR_EVENT2 events
PC Client firmware specifications for TPM 1.2 enabled platforms defined
an event log that mandated the use of SHA1 to hash event data and extend
the 20-byte digest into PCRs.

Version 2.0 of the specification enables platforms to hash event data
with one or more algorithms. If Platform Firmware extends digests to
PCRs using other hashing algorithms, an event in the event log has to
contain all of the recorded digests. This new log format is referred to
as a crypto agile log format.

Extend the library to support to support writing events for this format.

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: Id83ebe553b1485e9104a05a1ae814fd48ab2133b
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>