fix: dynamically determine digest prefix
We previously had constants for log prefix. This was inflexible and
prone to errors. The prefix is now dynamically determine for lines
following the first of the digest.

Add extra padding to the digest prefix so that the digest appears nested
under it's respective algorithm id in the print out. Previously this was
printed in-line, which causes issues with readability:

```
[NOTICE]     #2 AlgorithmId   : SHA512
[NOTICE]     Digest        : 3f 7f 88 ...
```

With the change:

```
[NOTICE]     #2 AlgorithmId   : SHA512
[NOTICE]        Digest        : 3f 7f 88 ...
```

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I5726d80e17ca879de68535d66692aab7483a5878