refactor!: standardise TCG_PCClientPCREvent write api
All entries in the event log except the first follow the
TCG_PCR_EVENT2.  The first event is special because it provides
information about identifies the version of the log, which implies the
format of the events, and the number and size of the recorded digests.

It is formatted as TCG_PCClientPCREvent structure, which is inherited
from the Conventional Spec (defined in Section 10.2.1
TCG_PCClientPCREvent Structure). Provide a helper for writing events
of this format for backward compatibility.

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: Ib302c124c1ce8f94db2e4592e27bb7176f2ff06d